backend done

+++ backend/controllers/product.controller.js
@@ -31,13 +31,17 @@ export const createProduct = async (req, res) => {
 
 export const deleteProduct = async (req, res) => {
     const {id} = req.params;
-    
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+        return res.status(404).json({ success: false, message: "Invalid Product Id"});
+    }    
+
     // let's add a response so that it will store in database
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Product Deleted" });
     } catch (error) {
-        res.status(404).json({ success: false, message: "Product not found"});
+        res.status(500).json({ success: false, message: "Server Error"});
     }
 };
